# docs/openapi/dpp-api-v1.yaml
openapi: 3.1.0
info:
  title: Claros DPP Public and Integration API
  version: "2026-06-27"
  description: >
    Current external API surface for Digital Product Passports. Public reads use
    `/api/public/passports/{dppId}` and return public fields by default. Selected
    restricted fields are returned only when a valid security group API key is
    supplied in `X-API-Key` or `X-Security-Group-Key`. Company automation writes
    use the company-name slug integration namespace and require a Bearer token.
servers:
  - url: https://www.claros-dpp.online
    description: Production
  - url: http://localhost:3001
    description: Local backend
...
paths:
...
  /api/public/passports/{dppId}/dynamic-values:
    get:
      tags: [Public read]
      summary: Read current dynamic field values
      description: >
        Returns public dynamic fields for a released or obsolete passport.
        A valid security group key also returns selected restricted dynamic
        fields covered by that key.
      parameters:
        - $ref: '#/components/parameters/DppIdPath'
        - $ref: '#/components/parameters/SecurityGroupKeyHeader'
        - $ref: '#/components/parameters/AlternateSecurityGroupKeyHeader'
      responses:
        '200':
          description: Latest allowed dynamic values keyed by canonical field key.
          content:
            application/json:
              schema:
                type: object
                required: [values]
                properties:
                  values:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        value: {}
                        updatedAt:
                          type: string
                          format: date-time
        '401':
          description: Invalid or revoked security group key.
        '403':
          description: Security group key is not valid for this company, type, or passport.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/public/passports/{dppId}/dynamic-values/{fieldKey}/history:
    get:
      tags: [Public read]
      summary: Read dynamic field history
      description: >
        Returns history for a public dynamic field. Restricted dynamic history
        requires a security group key that explicitly includes the field.
      parameters:
        - $ref: '#/components/parameters/DppIdPath'
        - name: fieldKey
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 500
        - $ref: '#/components/parameters/SecurityGroupKeyHeader'
        - $ref: '#/components/parameters/AlternateSecurityGroupKeyHeader'
      responses:
        '200':
          description: Ordered dynamic field history.
          content:
            application/json:
              schema:
                type: object
                required: [history]
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        value: {}
                        updatedAt:
                          type: string
                          format: date-time
        '401':
          description: Invalid or revoked security group key.
        '403':
          description: The field is restricted or the key does not cover it.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/public/passports/{dppId}/qrcode:
    get:
      tags: [Public read]
      summary: Read QR and carrier authenticity metadata
      parameters:
        - $ref: '#/components/parameters/DppIdPath'
      responses:
        '200':
          description: QR code and carrier authenticity metadata for a released or obsolete passport.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/companies/{companySlug}/integrations/v1/passports/{dppId}/dynamic-values:
    post:
      tags: [Integration write]
      summary: Push current dynamic field values for a passport
      description: Restricted dynamic fields are rejected even when the caller has a Bearer token.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CompanySlugPath'
        - $ref: '#/components/parameters/DppIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Dynamic values accepted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanySlugPath:
      name: companySlug
      in: path
      required: true
      schema:
        type: string
      description: Lowercase hyphenated company name, for example `acme-batteries`.
    DppIdPath:
      name: dppId
      in: path
      required: true
      schema:
        type: string
      description: Canonical DPP ID.
...
    SecurityGroupKeyHeader:
      name: X-API-Key
      in: header
      required: false
      schema:
        type: string
      description: Security group key for selected restricted fields.
    AlternateSecurityGroupKeyHeader:
      name: X-Security-Group-Key
      in: header
      required: false
      schema:
        type: string
      description: Alternate header for a security group key.
...
  responses:
    BadRequest:
      description: Invalid request.
    Unauthorised:
      description: Missing or invalid Bearer token.
    Forbidden:
      description: Authenticated caller is not allowed to perform this action.
    NotFound:
      description: Resource not found.
    RateLimited:
      description: Too many requests.
    InternalError:
      description: Internal server error.